% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{calculate_median}
\alias{calculate_median}
\title{Calculate Median}
\usage{
calculate_median(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The median of the vector.
}
\description{
This function calculates the median of a numeric vector.
}
\examples{
calculate_median(c(1, 2, 3, 4, 5))
}
