% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_dummy_matrix}
\alias{create_dummy_matrix}
\title{Create Dummy Variables for Categorical Predictors}
\usage{
create_dummy_matrix(X)
}
\arguments{
\item{X}{Matrix or data frame containing categorical variables.}
}
\value{
A matrix containing dummy variables.
}
\description{
Converts categorical variables within a matrix or data frame into dummy variables.
Each level of a factor results in a new binary column, except the first level to avoid multicollinearity.
}
\examples{
create_dummy_matrix(iris[, -1])
}
