% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_model.R
\name{ensemble_model_fitting}
\alias{ensemble_model_fitting}
\title{Ensemble Model Fitting Using Standardized Prediction}
\usage{
ensemble_model_fitting(data, response_var, model_list)
}
\arguments{
\item{data}{Data frame containing the training dataset including the response variable.}

\item{response_var}{Name of the response variable in the data frame.}

\item{model_list}{List of model function names as character strings. Each model is assumed to return an object
compatible with the predict_model function.}
}
\value{
A list containing combined predictions and details for each fitted model.
}
\description{
This function fits multiple models to the same dataset and combines their predictions using a
standardized prediction method. Supports various regression model types.
}
\examples{
\dontrun{
data(mtcars)
model_functions <- c("fit_ridge_model", "fit_logistic_model", "fit_random_forest_model")
result <- ensemble_model_fitting(mtcars, "mpg", model_functions)
print(result$combined_predictions)
}
}
