% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{fit_linear_internal}
\alias{fit_linear_internal}
\title{Internal function to fit linear model}
\usage{
fit_linear_internal(y, X, add_intercept)
}
\arguments{
\item{y}{Response variable vector.}

\item{X}{Design matrix with predictors.}
}
\value{
A list containing model fitting details.
}
\description{
Internal function to fit linear model
}
