% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model.R
\name{fit_log_internal}
\alias{fit_log_internal}
\title{Internal function to fit logistic regression model}
\usage{
fit_log_internal(X, y, add_intercept)
}
\arguments{
\item{X}{Design matrix with predictors.}

\item{y}{Response variable vector.}
}
\value{
A list containing model fitting details.
}
\description{
Internal function to fit logistic regression model
}
