% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest_model.R
\name{fit_random_forest_model}
\alias{fit_random_forest_model}
\title{Random Forest Model Fitting}
\usage{
fit_random_forest_model(y, X, model_type = "gaussian")
}
\arguments{
\item{y}{Response vector with continuous or binary outcomes.}

\item{X}{Data frame of predictors, either numeric or factor types.}

\item{model_type}{A character string specifying the type of model: 'gaussian' for regression or 'binomial' for classification.}
}
\value{
A list containing the random forest model object with model details and fitted values.
}
\description{
Fits a random forest model to the data, useful for both regression and classification.
Random forest is an ensemble learning method that operates by constructing multiple decision trees
at training time and outputting the class that is the mode of the classes (classification) or
mean prediction (regression) of the individual trees.
}
\examples{
data(mtcars)
result_rf <- fit_random_forest_model(mtcars$mpg, mtcars[, -1], model_type = 'gaussian')
}
