% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_model}
\alias{predict_model}
\title{Predict Method for Regression Models}
\usage{
predict_model(model, newdata, type = "response")
}
\arguments{
\item{model}{The model object returned by fitting functions.}

\item{newdata}{New data on which to predict responses.}

\item{type}{The type of prediction required: "response" for predicted values, "coefficients" for model coefficients
(not applicable for Random Forest), or "probabilities" for logistic regression probability estimates.}
}
\value{
Predicted values, coefficients, or probabilities based on the input model and specified type.
}
\description{
Generates predictions from regression models fitted using various techniques,
including standard, bagged approaches, and Random Forest. Handles linear, logistic, elastic net models,
and Random Forest.
}
