% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list containing the mean and median of the vector.
}
\description{
This function calculates the mean and median of a numeric vector.
}
\examples{
summarize_data(c(1, 2, 3, 4, 5))
}
