% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_model.R
\name{ensemble_model_fitting}
\alias{ensemble_model_fitting}
\title{Ensemble Model Fitting Using Standardized Prediction}
\usage{
ensemble_model_fitting(X, y, model_type, model_list, threshold = 0.5)
}
\arguments{
\item{X}{Data frame or matrix of predictors.}

\item{y}{Vector of response variables.}

\item{model_type}{Character specifying whether to fit models for 'gaussian' (regression) or 'binomial' (classification).}

\item{model_list}{List of model fitting functions expressed as character strings.}

\item{threshold}{Threshold for converting probabilities to class labels in case of 'binomial' model type, default is 0.5.}
}
\value{
A list containing combined predictions and details for each fitted model including individual model predictions.
}
\description{
Fits multiple models specified by the model list to the same dataset and combines their predictions using a
standardized prediction method. Supports regression (gaussian) and classification (binomial) types.
}
\examples{
  # Assuming `fit_linear_model` and `fit_random_forest_model` are used.
  data(mtcars)
  X <- mtcars[, -which(names(mtcars) == "mpg")]
  y <- mtcars$mpg  # For gaussian; use a binary response for 'binomial'
  model_list <- c("fit_linear_model", "fit_random_forest_model")
  results <- ensemble_model_fitting(X, y, model_type = 'gaussian', model_list = model_list)
  print(results$combined_predictions)
  # To view individual model details and predictions:
  print(results$model_details)
  print(results$individual_predictions)
}
