% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elastic_net_model.R
\name{fit_elastic_net_model}
\alias{fit_elastic_net_model}
\title{Elastic Net Regression Fitting}
\usage{
fit_elastic_net_model(
  y,
  X,
  alpha = 0.5,
  lambda = NULL,
  use_bagging = FALSE,
  R = 100
)
}
\arguments{
\item{y}{A response vector with continuous or binary outcomes.}

\item{X}{A matrix of predictors.}

\item{alpha}{The elastic net mixing parameter, with 0 <= alpha <= 1.}

\item{lambda}{The regularization penalty parameter for elastic net regression.
If NULL, cv.glmnet is used to find the optimal lambda.}
}
\value{
A glmnet model object fitted with elastic net regression.
}
\description{
Fits an elastic net regression model using glmnet to a given set of predictors and response.
}
\examples{
data(mtcars)
result_elastic_net <- fit_elastic_net_model(mtcars$mpg, mtcars[, c("hp", "wt")], alpha = 0.5)
}
