% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_model.R
\name{fit_lasso_model}
\alias{fit_lasso_model}
\title{Lasso Regression Fitting}
\usage{
fit_lasso_model(y, X, lambda = NULL, bagging = FALSE, R = 100)
}
\arguments{
\item{y}{A response vector with continuous or binary outcomes.}

\item{X}{A matrix of predictors.}

\item{lambda}{The regularization penalty parameter for lasso regression.
If NULL, cv.glmnet is used to find the optimal lambda.}
}
\value{
A glmnet model object fitted with lasso regression.
}
\description{
Fits a lasso regression model using glmnet to a given set of predictors and response.
}
\examples{
data(mtcars)
result_lasso <- fit_lasso_model(mtcars$mpg, mtcars[, c("hp", "wt")])
}
