% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_model.R
\name{fit_ridge_model}
\alias{fit_ridge_model}
\title{Ridge Regression Fitting}
\usage{
fit_ridge_model(y, X, lambda = NULL, bagging = FALSE, R = 100)
}
\arguments{
\item{y}{A response vector with continuous or binary outcomes.}

\item{X}{A matrix of predictors.}

\item{lambda}{The regularization penalty parameter for ridge regression.
If NULL, cv.glmnet is used to find the optimal lambda.}
}
\value{
A glmnet model object fitted with ridge regression.
}
\description{
Fits a ridge regression model using glmnet to a given set of predictors and response.
}
\examples{
data(mtcars)
result_ridge <- fit_ridge_model(mtcars$mpg, mtcars[, c("hp", "wt")])
}
