% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_model.R
\name{fit_ridge_model}
\alias{fit_ridge_model}
\title{Fit Ridge Regression Model with Optional Bagging and Intercept}
\usage{
fit_ridge_model(
  y,
  X,
  lambda = NULL,
  model_type = "gaussian",
  add_intercept = TRUE,
  bagging = FALSE,
  R = 100
)
}
\arguments{
\item{y}{Response vector with continuous or binary outcomes.}

\item{X}{Matrix of predictors, either numeric or factor types.}

\item{lambda}{Regularization penalty parameter for ridge regression; if NULL, it is determined via
cross-validation.}

\item{model_type}{A character string specifying whether to use 'gaussian' for regression or 'binomial' for classification.}

\item{add_intercept}{Logical indicating whether to include an intercept in the model.}

\item{bagging}{Logical indicating whether to perform bagging.}

\item{R}{Integer specifying the number of bootstrap samples to use if bagging is enabled.}
}
\value{
A list containing the ridge regression model details, or if bagging is enabled, an aggregated
result from multiple bootstrap samples including the model, lambda value, coefficients,
and fitted values (continuous values for gaussian and probability values for binomial)
}
\description{
Performs ridge regression using glmnet, optimizing the balance between model complexity
and prediction accuracy through L2 regularization. This method is ideal for handling multicollinearity,
reducing overfitting, and shrinking coefficient values, particularly in situations where the number of
predictors exceeds the number of observations. Optionally, bagging can be incorporated to further
enhance model stability and prediction accuracy by aggregating results from multiple bootstrap samples.
}
\examples{
# Example for Gaussian model
data(mtcars)
X <- mtcars[, c("hp", "wt")]
y <- mtcars$mpg
result_ridge_gaussian <- fit_ridge_model(y, X, model_type = "gaussian", add_intercept = TRUE, bagging = FALSE)
print(result_ridge_gaussian)

# Example for Binomial model
data(mtcars)
X <- mtcars[, c("hp", "wt")]
y <- ifelse(mtcars$am == 1, 1, 0)  # Converting 'am' to a binary outcome
result_ridge_binomial <- fit_ridge_model(y, X, model_type = "binomial", add_intercept = TRUE, bagging = TRUE, R = 50)
print(result_ridge_binomial)
}
